requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
mysqlclient
Pillow
supabase

# Channels / WebSocket support
channels
channels-redis

# Auth / docs / env
djangorestframework-simplejwt
drf-yasg
python-dotenv

# Optional helpers
requests
